fix(#44): move agent/workspace management out of core/
- Delete core/agents.py (superseded by services/agents/agent_registry.py)
- Delete tests/unit/core/test_agents.py (sole caller of deleted module)
- Move core/workspace_manifest.py → contracts/workspace_manifest.py (pure data model)
- Update 9 callers to import from nexus.contracts.workspace_manifest
