+++ src/nexus/contracts/workspace_manifest.py
