+++ src/nexus/services/context_branch.py
@@ -35,7 +35,7 @@ from nexus.contracts.exceptions import (
     NexusFileNotFoundError,
     StalePointerError,
 )
-from nexus.core.workspace_manifest import WorkspaceManifest
+from nexus.contracts.workspace_manifest import WorkspaceManifest
 from nexus.services.workspace_permissions import check_workspace_permission
 from nexus.storage.models import ContextBranchModel, WorkspaceSnapshotModel
