+++ src/nexus/services/workspace_manager.py
@@ -12,7 +12,7 @@ from typing import TYPE_CHECKING, Any
 from sqlalchemy import desc, select
 
 from nexus.contracts.exceptions import NexusFileNotFoundError
-from nexus.core.workspace_manifest import WorkspaceManifest
+from nexus.contracts.workspace_manifest import WorkspaceManifest
 from nexus.services.workspace_permissions import check_workspace_permission
 from nexus.storage.models import WorkspaceSnapshotModel
