+++ tests/e2e/self_contained/test_overlay_nexusfs.py
@@ -17,9 +17,9 @@ import pytest
 
 from nexus.backends.local import LocalBackend
 from nexus.contracts.exceptions import NexusFileNotFoundError
+from nexus.contracts.workspace_manifest import ManifestEntry, WorkspaceManifest
 from nexus.core.metadata import FileMetadata
 from nexus.core.metastore import MetastoreABC
-from nexus.core.workspace_manifest import ManifestEntry, WorkspaceManifest
 from nexus.services.overlay_resolver import OverlayResolver
