+++ tests/e2e/test_context_branch_e2e.py
@@ -25,8 +25,8 @@ from nexus.contracts.exceptions import (
     BranchStateError,
     NexusPermissionError,
 )
+from nexus.contracts.workspace_manifest import ManifestEntry, WorkspaceManifest
 from nexus.core.response import HandlerResponse
-from nexus.core.workspace_manifest import ManifestEntry, WorkspaceManifest
 from nexus.services.context_branch import ContextBranchService
 from nexus.storage.models._base import Base
 from nexus.storage.models.filesystem import WorkspaceSnapshotModel
